adding in language specific files, added the three go funcs
